Category crud operation working properly

+++ src/main/java/com/example/personaldiary/category/CategoryResource.java
@@ -1,5 +1,7 @@
 package com.example.personaldiary.category;
 
+import com.example.personaldiary.user.User;
+import com.example.personaldiary.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,6 +21,9 @@ public class CategoryResource {
 
     @Autowired
     private CategoryRepository categoryRepository;
+
+    @Autowired
+    private UserRepository userRepository;
 
     @GetMapping("/api/categories")
     public List<Category> getCategories() {
@@ -40,6 +45,11 @@ public class CategoryResource {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found (id-" + id + ")");
         }
 
+        UserDetails ud = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!category.get().getUser().getUsername().equals(ud.getUsername())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not allowed to get others data");
+        }
+
         return category.get();
     }
 
@@ -50,12 +60,25 @@ public class CategoryResource {
         if (!category.isPresent()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found (id-" + id + ")");
         }
+        if(category.get().getIsDefault()) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Default Categories can not be deleted");
+        }
+
+        UserDetails ud = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!category.get().getUser().getUsername().equals(ud.getUsername())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You can not delete others data");
+        }
 
         categoryRepository.delete(category.get());
     }
 
     @PostMapping("/api/category")
     public ResponseEntity<Object> createCategory(@RequestBody Category category) {
+        UserDetails ud = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User currentUser = userRepository.findByUsername(ud.getUsername());
+        category.setUser(currentUser);
+        category.setIsDefault(false);
+
         Category savedCategory = categoryRepository.save(category);
 
         URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
@@ -71,8 +94,17 @@ public class CategoryResource {
         if (!categoryOptional.isPresent()) {
             return ResponseEntity.notFound().build();
         }
+        if(categoryOptional.get().getUser() == null) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+        UserDetails ud = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!categoryOptional.get().getUser().getUsername().equals(ud.getUsername())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         category.setId(id);
+        category.setUser(categoryOptional.get().getUser());
+        category.setIsDefault(false);
         categoryRepository.save(category);
 
         return ResponseEntity.noContent().build();
